preparation for esp_idf

+++ components/ehmtxv2/EHMTX.cpp
@@ -706,7 +706,7 @@ namespace esphome
         }
       }
       this->display_gauge = true;
-      ESP_LOGD(TAG, "show_gauge 2 color %d", round(percent));
+      ESP_LOGD(TAG, "show_gauge 2 color %d", int(round(percent)));
     }
   }
 #else
@@ -720,7 +720,7 @@ namespace esphome
       this->display_gauge = true;
       this->gauge_value = (uint8_t)(100 - percent) * 7 / 100;
     }
-    ESP_LOGD(TAG, "show_gauge 2 color %d", round(percent));
+    ESP_LOGD(TAG, "show_gauge 2 color %d", int(round(percent)));
   }
 #endif
